+++ adagucserverEC/GenericDataWarper/GDWDrawFunctionSettings.cpp
@@ -71,7 +71,7 @@ GDWDrawFunctionSettings getDrawFunctionSettings(CDataSource *dataSource, CDrawIm
       } else {
         settings.intervals.reserve(numShadeDefs);
         for (const auto &shadeInterval : styleConfiguration->shadeIntervals) {
-          settings.intervals.push_back(Interval({.min = atof(shadeInterval.attr.min.c_str()), .max = atof(shadeInterval.attr.max.c_str()), .color = CColor(shadeInterval.attr.fillcolor.c_str())}));
+          settings.intervals.push_back(Interval({.min = shadeInterval.attr.min, .max = shadeInterval.attr.max, .color = CColor(shadeInterval.attr.fillcolor.c_str())}));
         }
         // Sort shaded intervals on min value
         std::sort(settings.intervals.begin(), settings.intervals.end(), [](const Interval &left, const Interval &right) { return left.min < right.min; });